openai_agents_sdk_mcp_integration/02_tool_filters/pyproject.toml
[project]
name = "02-tool-filters"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.13.1",
    "openai-agents==0.2.6",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
